готовим рефактор по замечаниям
