+++ src/lesson3/__init__.py
@@ -21,4 +21,27 @@
 Для поворачивающегося объекта определен интерфейс, устойчивый к появлению
     новых видов движущихся объектов - 1 балл
 Реализован тесты - 4 балла (по одному баллу за тест, но не более 4-х баллов).
+
+Итерация 2
+===========
+
+интерфейсы
+-----------
+
+- Команда GenericCommand
+- Объект, который умеет двигаться по прямой = MovableInterface
+- Объект, который умеет поворачиваться = SpaceDirectionInterface
+- UObject = DynamicInterfaceObject. Не совсем интерфейс, но очень близко.
+- Убираем PositionedInterface
+Я думаю, что нужно оставить SpaceVectorInterface, чтобы не привязваться
+к конкретной реализации вектора (2D, 3D, декартов, один из полярных...)
+
+классы:
+----------
+
+- Команда Move = StraightMoveCommand
+- Команда Rotate = RotationCommand
+- Вектор = SpaceVector2
+- Убираем Position Movable SpaceDirection
+
 """
